service-workers/service-worker/fetch-filter-escalation.https.html

<!DOCTYPE html>
<title>Service Worker: fetch filter escalation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script>
var host_info = get_host_info();
promise_test(function(t) {
    var SCOPE = './resources/';
    var SCRIPT = 'resources/fetch-filter-escalation-worker.js';

    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          add_completion_callback(function() { frame.remove(); });
		  window.iFetch = frame.contentWindow.fetch;
        })
	  .then(defineTests)
      .catch(unreached_rejection(t))
	  .then(done);
}, 'setup');

function defineTests() {
  promise_test(function(t) {
    return fetch(host_info.HTTPS_ORIGIN + '/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=*')
	  .then(function(response) {
		  assert_true(response.headers.has('x-serviceworker-serverheader'));
	    });
  }, 'Local origin has extra header');

  promise_test(function(t) {
    return fetch(host_info.HTTPS_REMOTE_ORIGIN + '/service-workers/service-worker/resources/fetch-access-control.py?ACAOrigin=*')
	  .then(function(response) {
		  assert_false(response.headers.has('x-serviceworker-serverheader'));
	    });
  }, 'Remote origin does not have extra header');

  promise_test(function(t) {
    return iFetch(host_info.HTTPS_ORIGIN + '/service-workers/service-worker/resources/proxy')
	  .then(function(response) {
		  assert_true(response.headers.has('x-serviceworker-serverheader'));
	    });
  }, 'Local origin through Service Worker');

  promise_test(function(t) {
    return iFetch(host_info.HTTPS_REMOTE_ORIGIN + '/service-workers/service-worker/resources/proxy')
	  .then(function(response) {
		  assert_false(response.headers.has('x-serviceworker-serverheader'));
	    });
  }, 'Remote origin through Service Worker');
}
</script>
